+++ src/OpenTelemetry.Api/Logs/LoggerProvider.cs
@@ -67,16 +67,33 @@ public class LoggerProvider : BaseProvider
     internal
 #endif
         Logger GetLogger(string? name, string? version)
+        => this.GetLogger(new LoggerOptions() { Name = name, Version = version });
+
+#if EXPOSE_EXPERIMENTAL_FEATURES
+    /// <summary>
+    /// Gets a logger with the given name and version.
+    /// </summary>
+    /// <remarks><inheritdoc cref="Logger" path="/remarks"/></remarks>
+    /// <param name="options">The options to use to create the logger.</param>
+    /// <returns><see cref="Logger"/> instance.</returns>
+    [Experimental(DiagnosticDefinitions.LogsBridgeExperimentalApi, UrlFormat = DiagnosticDefinitions.ExperimentalApiUrlFormat)]
+    public
+#else
+    internal
+#endif
+        Logger GetLogger(LoggerOptions options)
     {
-        if (!this.TryCreateLogger(name, out var logger))
+        Guard.ThrowIfNull(options);
+
+        if (!this.TryCreateLogger(options, out var logger))
         {
             return NoopLogger;
         }
 
 #if NET
-        logger.SetInstrumentationScope(version);
+        logger.SetInstrumentationScope(options?.Version, options?.SchemaUrl);
 #else
-        logger!.SetInstrumentationScope(version);
+        logger!.SetInstrumentationScope(options?.Version, options?.SchemaUrl);
 #endif
 
         return logger;
@@ -87,8 +104,8 @@ public class LoggerProvider : BaseProvider
     /// Try to create a logger with the given name.
     /// </summary>
     /// <remarks><inheritdoc cref="Logger" path="/remarks"/></remarks>
-    /// <param name="name">Optional name identifying the instrumentation library.</param>
-    /// <param name="logger"><see cref="Logger"/>.</param>
+    /// <param name="options">The options to use to create the logger.</param>
+    /// <param name="logger">If successful, contains the created <see cref="Logger"/>.</param>
     /// <returns><see langword="true"/> if the logger was created.</returns>
     [Experimental(DiagnosticDefinitions.LogsBridgeExperimentalApi, UrlFormat = DiagnosticDefinitions.ExperimentalApiUrlFormat)]
     protected
@@ -96,7 +113,7 @@ public class LoggerProvider : BaseProvider
     internal
 #endif
         virtual bool TryCreateLogger(
-        string? name,
+        LoggerOptions options,
 #if NETSTANDARD2_1_OR_GREATER || NET
         [NotNullWhen(true)]
 #endif